Move UIColor(hexRGB:) to its own extension file, added to build target
Define the hex init in Extensions/UIColorExtension.swift so it's
unambiguously available to all files. Registers the file in the
Xcode project (PBXBuildFile, PBXFileReference, group, Sources phase).
