+++ Go Cycling/View Model/UserPreferences.swift
@@ -38,18 +38,3 @@ class UserPreferences: ObservableObject {
     }
 
 }
-
-extension UIColor {
-    convenience init?(hexRGB hex: String) {
-        let cleaned = hex.trimmingCharacters(in: CharacterSet.alphanumerics.inverted)
-        guard cleaned.count == 6 else { return nil }
-        var value: UInt64 = 0
-        Scanner(string: cleaned).scanHexInt64(&value)
-        self.init(
-            red:   CGFloat((value >> 16) & 0xFF) / 255,
-            green: CGFloat((value >>  8) & 0xFF) / 255,
-            blue:  CGFloat( value        & 0xFF) / 255,
-            alpha: 1
-        )
-    }
-}
